# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MMO Battle Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /player:
    get:
      operationId: getPlayer
      tags: [player]
      summary: Get player info
      responses:
        "200":
          description: Player data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: No player yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createPlayer
      tags: [player]
      summary: Create or reset a player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlayerRequest"
      responses:
        "200":
          description: Player created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
...
  /player/ascend:
    post:
      operationId: ascendPlayer
      tags: [player]
      summary: Ascend — reset level to 1, gain a permanent ascension bonus
      responses:
        "200":
          description: Ascension successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AscendResponse"
        "400":
          description: Cannot ascend yet (level too low or battle active)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    CreatePlayerRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    Player:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        level:
          type: integer
        xp:
          type: integer
        xpToNextLevel:
          type: integer
        hp:
          type: integer
        maxHp:
          type: integer
        attack:
          type: integer
        defense:
          type: integer
        gold:
          type: integer
        monstersDefeated:
          type: integer
        enchantingStones:
          type: integer
        meleeSkillLevel:
          type: integer
        meleeSkillXp:
          type: integer
        meleeSkillXpToNext:
          type: integer
        defenseSkillLevel:
          type: integer
        defenseSkillXp:
          type: integer
        defenseSkillXpToNext:
          type: integer
        ascensionLevel:
          type: integer
      required:
        - id
        - name
        - level
        - xp
        - xpToNextLevel
        - hp
        - maxHp
        - attack
        - defense
        - gold
        - monstersDefeated
        - enchantingStones
        - meleeSkillLevel
        - meleeSkillXp
        - meleeSkillXpToNext
        - defenseSkillLevel
        - defenseSkillXp
        - defenseSkillXpToNext
        - ascensionLevel
...
    AscendResponse:
      type: object
      properties:
        player:
          $ref: "#/components/schemas/Player"
        message:
          type: string
        ascensionLevel:
          type: integer
      required:
        - player
        - message
        - ascensionLevel
...